+++ packages/sdk/src/loadProject.test.ts
@@ -16,13 +16,15 @@ import {
 	ProjectSettingsFileNotFoundError,
 	ProjectSettingsInvalidError,
 } from "./errors.js"
-import { createNodeishMemoryFs, normalizePath } from "@lix-js/fs"
+import { normalizePath } from "@lix-js/fs"
 import { createMessage } from "./test-utilities/createMessage.js"
 import { tryCatch } from "@inlang/result"
 import { mockRepo } from "@lix-js/client"
 import { type Snapshot } from "@lix-js/fs"
 // eslint-disable-next-line no-restricted-imports -- test
 import { readFileSync } from "node:fs"
+// eslint-disable-next-line no-restricted-imports -- test
+import { resolve } from "node:path"
 
 // ------------------------------------------------------------------------------------------------
 
@@ -109,6 +111,12 @@ const _import: ImportFunction = async (name) =>
 		default: name === "plugin.js" ? mockPlugin : mockMessageLintRule,
 	} satisfies InlangModule)
 
+const ciTestRepoSnapshot = JSON.parse(
+	readFileSync(resolve(__dirname, "../mocks/ci-test-repo-no-shallow.json"), {
+		encoding: "utf-8",
+	})
+) as Snapshot
+
 // ------------------------------------------------------------------------------------------------
 
 /**
@@ -118,11 +126,11 @@ const _import: ImportFunction = async (name) =>
  * like files: they can be renamed and moved around.
  */
 it("should throw if a project (path) does not have a name", async () => {
-	const fs = createNodeishMemoryFs()
+	const repo = await mockRepo()
 	const project = await tryCatch(() =>
 		loadProject({
 			projectPath: "/source-code/.inlang",
-			nodeishFs: fs,
+			repo,
 			_import,
 		})
 	)
@@ -130,7 +138,7 @@ it("should throw if a project (path) does not have a name", async () => {
 })
 
 it("should throw if a project path does not end with .inlang", async () => {
-	const fs = createNodeishMemoryFs()
+	const repo = await mockRepo()
 
 	const invalidPaths = [
 		"/source-code/frontend.inlang/settings",
@@ -142,7 +150,7 @@ it("should throw if a project path does not end with .inlang", async () => {
 		const project = await tryCatch(() =>
 			loadProject({
 				projectPath: invalidPath,
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 		)
@@ -152,12 +160,12 @@ it("should throw if a project path does not end with .inlang", async () => {
 
 describe("initialization", () => {
 	it("should throw if projectPath is not an absolute path", async () => {
-		const fs = createNodeishMemoryFs()
+		const repo = await mockRepo()
 
 		const result = await tryCatch(() =>
 			loadProject({
 				projectPath: "relative/path",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 		)
@@ -166,10 +174,7 @@ describe("initialization", () => {
 	})
 
 	it("should generate projectId on missing projectid", async () => {
-		const ciTestRepo: Snapshot = JSON.parse(
-			readFileSync("./mocks/ci-test-repo-no-shallow.json", { encoding: "utf-8" })
-		)
-		const repo = await mockRepo({ fromSnapshot: ciTestRepo })
+		const repo = await mockRepo({ fromSnapshot: ciTestRepoSnapshot })
 
 		const existing = await repo.nodeishFs
 			.readFile("/project.inlang/project_id", {
@@ -205,10 +210,7 @@ describe("initialization", () => {
 	})
 
 	it("should reuse projectId on existing projectid", async () => {
-		const ciTestRepo: Snapshot = JSON.parse(
-			readFileSync("./mocks/ci-test-repo-no-shallow.json", { encoding: "utf-8" })
-		)
-		const repo = await mockRepo({ fromSnapshot: ciTestRepo })
+		const repo = await mockRepo({ fromSnapshot: ciTestRepoSnapshot })
 
 		repo.nodeishFs.writeFile("/project.inlang/project_id", "testId")
 
@@ -235,14 +237,15 @@ describe("initialization", () => {
 	})
 
 	it("should resolve from a windows path", async () => {
-		const fs = createNodeishMemoryFs()
+		const repo = await mockRepo()
+		const fs = repo.nodeishFs
 		fs.mkdir("C:\\Users\\user\\project.inlang", { recursive: true })
 		fs.writeFile("C:\\Users\\user\\project.inlang\\settings.json", JSON.stringify(settings))
 
 		const result = await tryCatch(() =>
 			loadProject({
 				projectPath: "C:\\Users\\user\\project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 		)
@@ -253,12 +256,13 @@ describe("initialization", () => {
 
 	describe("settings", () => {
 		it("should return an error if settings file is not found", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			fs.mkdir("/user/project", { recursive: true })
 
 			const project = await loadProject({
 				projectPath: "/user/non-existend-project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -266,13 +270,14 @@ describe("initialization", () => {
 		})
 
 		it("should return an error if settings file is not a valid JSON", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", "invalid json")
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -280,13 +285,14 @@ describe("initialization", () => {
 		})
 
 		it("should return an error if settings file is does not match schema", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify({}))
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -294,12 +300,13 @@ describe("initialization", () => {
 		})
 
 		it("should return the parsed settings", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -307,14 +314,15 @@ describe("initialization", () => {
 		})
 
 		it("should not re-write the settings to disk when initializing", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			const settingsWithDeifferentFormatting = JSON.stringify(settings, undefined, 4)
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", settingsWithDeifferentFormatting)
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -341,13 +349,14 @@ describe("initialization", () => {
 					default: {} as Plugin,
 				} satisfies InlangModule)
 
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import: $badImport,
 			})
 
@@ -374,12 +383,13 @@ describe("initialization", () => {
 describe("functionality", () => {
 	describe("settings", () => {
 		it("should return the settings", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -387,12 +397,13 @@ describe("functionality", () => {
 		})
 
 		it("should set a new settings", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -413,12 +424,13 @@ describe("functionality", () => {
 
 	describe("setSettings", () => {
 		it("should fail if settings are not valid", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -428,7 +440,8 @@ describe("functionality", () => {
 		})
 
 		it("should throw an error if sourceLanguageTag is not in languageTags", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 
 			const settings: ProjectSettings = {
@@ -441,7 +454,7 @@ describe("functionality", () => {
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -450,12 +463,13 @@ describe("functionality", () => {
 		})
 
 		it("should write settings to disk", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -477,7 +491,8 @@ describe("functionality", () => {
 
 	describe("installed", () => {
 		it("should return the installed items", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			const settings: ProjectSettings = {
 				sourceLanguageTag: "en",
 				languageTags: ["en"],
@@ -487,7 +502,7 @@ describe("functionality", () => {
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -508,7 +523,8 @@ describe("functionality", () => {
 		})
 
 		it("should apply 'warning' as default lint level to lint rules that have no lint level defined in the settings", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 
 			const settings: ProjectSettings = {
 				sourceLanguageTag: "en",
@@ -521,7 +537,7 @@ describe("functionality", () => {
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -550,7 +566,8 @@ describe("functionality", () => {
 				loadMessages: () => [{ id: "some-message", alias: {}, selectors: [], variants: [] }],
 				saveMessages: () => undefined,
 			}
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile(
 				"/user/project.inlang/settings.json",
@@ -568,7 +585,7 @@ describe("functionality", () => {
 			}
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -601,7 +618,8 @@ describe("functionality", () => {
 				loadMessages: () => [{ id: "some-message", alias: {}, selectors: [], variants: [] }],
 				saveMessages: () => undefined,
 			}
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile(
 				"/user/project.inlang/settings.json",
@@ -619,7 +637,7 @@ describe("functionality", () => {
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -633,12 +651,13 @@ describe("functionality", () => {
 
 	describe("errors", () => {
 		it("should return the errors", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 			project.errors.subscribe((errors) => {
@@ -649,12 +668,13 @@ describe("functionality", () => {
 
 	describe("customApi", () => {
 		it("should return the app specific api", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -666,12 +686,13 @@ describe("functionality", () => {
 
 	describe("messages", () => {
 		it("should return the messages", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -681,7 +702,8 @@ describe("functionality", () => {
 
 	describe("query", () => {
 		it("should call saveMessages() on updates", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 
 			const settings: ProjectSettings = {
 				sourceLanguageTag: "en",
@@ -716,7 +738,7 @@ describe("functionality", () => {
 
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -860,7 +882,8 @@ describe("functionality", () => {
 		 *  - Might be slow for a large number of messages. The requirement hasn't popped up yet though.
 		 */
 		it("should pass all messages, regardless of which message changed, to saveMessages()", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 
 			const settings: ProjectSettings = {
 				sourceLanguageTag: "en",
@@ -897,7 +920,7 @@ describe("functionality", () => {
 
 			const project = await loadProject({
 				projectPath: "/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 
@@ -912,12 +935,13 @@ describe("functionality", () => {
 
 	describe("lint", () => {
 		it.todo("should throw if lint reports are not initialized yet", async () => {
-			const fs = await createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project", { recursive: true })
 			await fs.writeFile("/user/project/project.inlang.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project/project.inlang.json",
-				nodeishFs: fs,
+				repo,
 				_import,
 			})
 			// TODO: test with real lint rules
@@ -934,12 +958,13 @@ describe("functionality", () => {
 				languageTags: ["en"],
 				modules: ["lintRule.js"],
 			}
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 			await fs.mkdir("/user/project.inlang", { recursive: true })
 			await fs.writeFile("/user/project.inlang/settings.json", JSON.stringify(settings))
 			const project = await loadProject({
 				projectPath: "/user/project.inlang",
-				nodeishFs: fs,
+				repo,
 				_import: async () => ({
 					default: mockMessageLintRule,
 				}),
@@ -951,7 +976,8 @@ describe("functionality", () => {
 
 	describe("watcher", () => {
 		it("changing files in resources should trigger callback of message query", async () => {
-			const fs = createNodeishMemoryFs()
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
 
 			const messages = {
 				$schema: "https://inlang.com/schema/inlang-message-format",
@@ -1006,7 +1032,7 @@ describe("functionality", () => {
 			// establish watcher
 			const project = await loadProject({
 				projectPath: normalizePath("/project.inlang"),
-				nodeishFs: fs,
+				repo,
 				_import: async () => ({
 					default: mockMessageFormatPlugin,
 				}),